+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Wrap basemap names in switcher.
 - Added header facet widget.
 - Added facet class to form page.
+- Block click actions on disabled buttons.
 
 # [0.3.1] - 2018-06-08
 - Prevent form submission if there are no fields in the page.